Update patch set 2
Patch Set 2:

I don't believe I've updated the Linux builders, so I think this would still break many things.

But I've been (the past few days) working on converting the Linux Docker builders to be buildlet-based VMs.  I have a tool that converts Docker images into GCE VMs.  So I'm about to change things.

I'd prefer we wait on this another week or so. But if you want to submit it for laughs and see what breaks, feel free, though I'll be traveling in 36 hours for FOSDEM.

Patch-set: 2