Update patch set 3
Patch Set 3: Code-Review+2

I am not convinced this is correct, but I downloaded it and built it and tried a bunch of cases I thought would break, and they didn't. So in the absence of a concrete proof that it's wrong, it can go in. But if anyone sees escape analysis bugs, this is the first place to look.

To be clear, the problem is not the new code per se, but the fact that I didn't think the algorithm here was powerful enough to model this. But it does seem to be working.

There may be more work on escape analysis later this year anyway.

Patch-set: 3
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2