Update patch set 1
Patch Set 1: Code-Review-1

It is a bit difficult to infer from the changes here, but it looks like the problem you are solving is that s390x puts the data at a very high address, triggering the panic about dummy being too far out.

It looks like s390 (non-x) puts the data only at 4 MB, which would work fine with the existing test.

In that case, rather than introduce significant complexity in the test here, I think it would be fine to put

// +build !s390x

at the top of the original file, disabling the test on s390x instead of trying to write one that works. 

Practically speaking, the nil checks are emitted in the portable half of the compiler. If they are emitted everywhere but s390x, they're going to be emitted on s390x too.

Patch-set: 1
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1