Update patch set 8
Patch Set 8:

We don't combine files for other OSes, so I'd like to keep the files as
is (the only exception is ppc64 and ppc64le, but that's only a difference
in endianness)

Regarding the frozen syscall package, I think we've decided that new
OS port should provide the same interface as already supported OSes
provide when the ppc64 port went in. Just providing the minimum
interface will make using syscall package significantly harder even for
the same OS. Think if a developer tried to move a working package
that uses syscall pakage from Darwin/386 to Darwin/arm. Ideally the
package should x/sys sub-repo, but the original package works fine
with the syscall package why bother changing it?

IMHO, the frozen status of the syscall package is to prevent adding
new syscalls and structures to the package for existing platforms.

Patch-set: 8
Label: Code-Review=0