Update patch set 3
Patch Set 3:

Ready for review.

The code still contains potential nil deref that Totoro described here:
https://go-review.googlesource.com/#/c/2103
But the crash does not happen in my tests. So I want to get the high-level structure first and get all the pieces in, and then work on improvements and fixes in smaller incremental changes.

Patch-set: 3
Label: Code-Review=0