test: use +test tag for run.go (instead of +build)
This is an alternative to http://golang.org/cl/3675.

The main issue is that the misc/cgo/{stdio,life} tests are
silently getting skipped when invoked from run.bash.

The main source of confusion is that both "go build"
and "test/run.go" try to interpret the same tag (+build).
They don't have to. Using a different tag (+test) for
run.go resolves this confusion and makes it clear which
tool the tag comments are intended for.

Change-Id: Ib73e1ecb6a6deea03b86995560966023ace93c25