net/http: allow Hijack to be used with CloseNotify
The existing implementation makes it impossible to run a Server that
both gets notified about disconnecting clients and is capable of
handling upgrade requests (such as to websocket protocol).

This change allows connections that have been used with CloseNotify to
also be used with Hijack. It does so by returning a wrapped net.Conn
type that reads from the closeNotify pipe reader, rather than directly
from net.Conn.

Change-Id: Ic250dc8dbdee3e3063846b1b3b0b7b21939b1529