Update patch set 3
Patch Set 3:

> Needs a new test.

I've added a test that calls CloseNotify() and then immediately hijacks the conn.

It seemed like I should verify that the hijacked conn still works in this case, so the test has a lot of duplication with TestWriteAfterHijack. Let me know if any of that is unnecessary.

Also, I opened #9763 to track this. We've got one other possible solution proposed there.

Patch-set: 3