Update patch set 3
Patch Set 3:

This is not right, or at least not the way I'd like to fix this problem.
And it doesn't seem to have fixed the build.
Given that it didn't fix anything and I want to do it differently, reverting.

Patch-set: 3
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0