Update patch set 6
Patch Set 6: Code-Review-1

I'm somewhat nervous to design a synchronization primitive which behaves poorly in non-obvious ways when used concurrently in apparently innocent ways. In the described scenario, G3 simply reused the same WaitGroup value after explicitly waiting for it to be unblocked under "G3: Wait()". To a developer, this feels like "alright, everyone else is done and this wg is ready for further use". If that happens, the outcome is a rare and obscure bug which will be hard to trace back into a misuse of the WaitGroup implementation. We could proudly say "We told you that was not okay", but we probably won't have to, as they'll never know.

Patch-set: 6
Reviewer: Gerrit User 5666 <5666@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1