Update patch set 6
Patch Set 6:

> Would it be clearer to say
> "sync: WaitGroup misuse: Add called concurrently with Wait"

I haven't looked at the details of this PR but
I'm slightly concerned by comments like this.

There are circumstances where I believe it should be fine
to call Add concurrently with Wait, and I have
code that relies on this.

For example http://play.golang.org/p/5dF4oMIqmi

This is not an uncommon scenario (one worker
spawns a bunch of others).

I'm hoping to be reassured that this is just fine
and won't be affected at all by the changes in this
CL...

Patch-set: 6
Label: Code-Review=0