Update patch set 2
Patch Set 2:

> Sorry for dragging this out, but what if I type make(chan int, 0) ?
 > Isn't that also an ideal 0?

The 0 is ideal 0 until it gets to the typechecking for TCHAN/TMAP and converts all arguments into TINT. If there are no arguments, it explicitly inserts the ideal int 0 created by nodintconst(0) (https://github.com/golang/go/blob/master/src/cmd/gc/typecheck.c#L1561). It's probably not correct to rely on this behavior, but I'm not aware for a way to check for nodintconst(0) other than the fact that it is TIDEAL.

FWIW, I do handle that case in the added test.

Patch-set: 2
Label: Code-Review=0