Update patch set 1
Patch Set 1: Code-Review-1

Sorry but no. I don't want to touch this code right now, and this is not an improvement.

FWIW, I am checking that the entire tree still builds cross-compiling for Plan 9 and that the same bits come out before and after my various changes, so Plan 9 should not be too far behind, and if this change makes all.bash work for you, then it will get fixed for free by the conversion.

Patch-set: 1
Label: Code-Review=-1