cmd/ld: always mark tlsg STLSBSS
Android apps build again.

Defining TLSG in runtime/tls_arm.s gives it the type SNOPTRBSS, so its
type was never being set when GOOS=android. I considered modifying the
if statement, but I no longer understand the intention of the original
change (in d738c6b0cad). We were always setting it before, what
platform is this not valid for?

Fixes #9829

Defining

Change-Id: I3eaa4a9590893eff67695797eb22547a170cdbcd