Update patch set 2
Patch Set 2:

> > (1 comment)
 > >
 > > i think the implementation is ad-hoc, ideally we should
 > > leverage the existing action mechanism. leave for others
 > > to decide.
 > 
 > I agree.
 > This change makes build order non-deterministic even with -p=1.
 > Also the new tasks completely ignore priorities, and there is a
 > pending issue for critical-path scheduling.
 > Also I was thinking about automatic test sharding, that will also
 > require dynamic emission of actions.

Where is the non-determinism with -p=1? It is more than possible that I'm missing something obvious. With -p=1 there will only be a single goroutine created in builder.do() and that goroutine will call down into builder.cgo(). The order of tasks added to to the tasks slice in builder.cgo() is exactly the same as the old code and they will be executed in the same order if there is only a single goroutine running them. 

Dynamic emission of actions seems a bit more invasive, though doable. I think the actions for individual compilations would need to be given the priority of the package action and thus get placed at the front of the ready queue. I can certainly explore that if that direction is preferred.

Patch-set: 2