Update patch set 2
Patch Set 2:

Currently the behavior is such that Eq returns true iff both points from both Rectangle's are strictly equal. Since Rectangle and Point are both concrete types covered under the compatibility guarantee, `r == s` for any two Rectangle will continue to work.

Since the behavior of Eq is not strictly defined in the documentation, it may be reasonable to overload to set it apart from language operator equality. Would that be a violation of the guarantee though? Such a change may break some programs that depend on the observed behavior of passing empty rectangles.

I can imagine cases where some applications would want to treat an rectangle as if it were a "fat" Point; that assumption proved true prior to this change. If such cases are prevalent, it would just be easy enough for new applications to check for empty-sensitive equality with any of (depending on the common case being optimized for):

r.Empty() && s.Empty() || r.Eq(s)
r.Empty() && s.Empty() || r == s
r == s || r.Empty() && s.Empty()

Patch-set: 2