syscall: Add Foreground and Joinpgrp to SysProcAttr
On Unix, when placing a child in a new process group allow that group to
become the foreground process group. Also, allow a child process to join
a specific process group.

When setting the foreground process group Ctty is used as the file
descriptor of the controlling terminal. This has been added to the BSD
and Solaris SysProcAttr structures and the handling of Setctty changed
to match Linux.

Change-Id: I18d169a6c5ab8a6a90708c4ff52eb4aded50bc8c