Update patch set 2
Patch Set 2:

> (Just saw CL description, didn't read code.) I did this once
 > before. I decided against it because it is essentially a backdoor
 > language change. I still believe that's true. I'm not convinced
 > this is a good idea.

Do you mean that it speedups programs too much? But isn't it true for all serious performance optimizations like continuous stacks, concurrent GC, precise GC, capturing by value? Take precise GC, now people don't zero dead local vars, and that can break with Go1.0 and gccgo.
If we take this position, we won't be able to move forward. I though that our position is that we move forward, and if there is any issues in older releases the solution is to upgrade.
And this equally applies to bug fixes, whenever we fix a bug that is present in older releases, that's effectively a backdoor language change?

Patch-set: 2