Update patch set 2
Patch Set 2:

> An issue I have with this is that it essentially penalizes people
 > that take the care to stop the timer, and it's quite likely to be
 > that code that would really benefit from this.

This change does not penalize that case. NewTimer/Stop can be optimized to some degree as well. But that is orthogonal optimization. When user does NewTimer/Stop, there are chances that he executes select in a loop, and so it can make sense to indeed create and arm the timer only once, instead of N times on every iteration.
So if you execute select once, you do case <-time.After which is shorter, simpler, less error prone and faster as well. If you execute select in a loop, then create timer outside of the loop, and we will do this case as fast as possible as well.

Patch-set: 2
Reviewer: Gerrit User 5400 <5400@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0