Update patch set 2
Patch Set 2:

I see that in some cases it removed $ at the end of error messages. With this thing we can add ^$ to all messages. Not sure whether it's worth it.

Also, it replaced "autotmp_[0-9]+" and "autotmp*" with actual temp numbers. Need to handle it in the updater.

Patch-set: 2
Label: Code-Review=0