Update patch set 1
Patch Set 1:

I do not know myself but that is not a topic for this patch.  The goal of this patch is to resolves the type on-demand, only when it is really needed.  This prevents breaking runtime-gdb.py load when something changes.

Patch-set: 1
Reviewer: Gerrit User 6760 <6760@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0