syscall: Reimplement linux syscalls in terms of their *at replacements.
Updates #9974

NOT READY FOR SUBMISSION YET

This proposal tackles the body of syscalls which have been replaced, and are
now deprecated in linux. This is needed for the arm64 port as arm64 is the
first linux architecture to remove the "legacy" forms of these syscalls.

The *AT variants were added in kernel 2.6.16, so well before our 2.6.23 cutoff
(hey, it'll even work on RHEL5).

Discussion: https://groups.google.com/forum/#!topic/golang-dev/zpeFtN2z5Fc

Change-Id: I473a7c9a295d6f776fcdc75dcce06cbe9e3564ee