Update patch set 11
Patch Set 11: Code-Review+1

The code looks correct and good to me.

I have a concern with block size and alignment. Runtime sometime asks for blocks that are not multiple of page size. First problem here is that after s block split, sysAlloc can return a block that does not start at multiple of page size (aligned by just 8 or so). Second problem here is that it can lead to excessive internal fragmentation. What do you think if we round up size to _PhysPageSize in the beginning of both sysAlloc and sysFree? It seems that it is the only modification that is required to fix both problems.

Also please add
const memDebug = false
and check invariants after all operations under if (that the list is ordered, there are no overlapping blocks, sysAlloc returns completely zeroed memory, etc). And do a run of all.bash with memDebug = true. Such kind of code is a sweet spot for offbyones.

Patch-set: 11
Label: Code-Review=+1