Update patch set 3
Patch Set 3:

> > You can disable inlining, but there is nothing really like -O0.
 > 
 > Isn't -N and -l the direct counterpart of -O0?

They are as close as we come, but I don't think they are the same.  -O0 gives you code in which each statement is separate and you can always ask for the value of any variable.  As far as I know gc does not really support that mode.

Patch-set: 3