compress/zlib: remap testdata for darwin/arm
This is a trial balloon CL, I'm looking for opinions.

The problem is darwin/arm takes a set of files and builds a bundle. The
binary goes in the root of the bundle, and runtime/cgo has been
modified to set the current working directory to the root. But the
binary being in the root means "../testdata" and accessing files from
other parts of the file system does not work. So I intend to remap
those packages that access testdata from their parent directories:

	compress/flate
	compress/lzw
	compress/zlib
	image/draw
	image/gif
	image/jpeg

and then have go_darwin_arm_exec copy any parent testdata directories
into the bundle.

An alternative to this is to copy the entire GOROOT into the bundle,
and then modify the test harness generated by the Go tool to change
the working directory. I'm worried about the performance hit of that,
it already takes ~3s to start each test by deploying an app.

Change-Id: I5baab5e204bb437680be23c7f5e4af778e60b932