Update patch set 2
Patch Set 2:

> > > I wonder if we should do something like nacl does for its
 > > > filesystem necessary for testing?
 > >
 > > Essentially nacl embeds everything into the syscall package.
 > > Perhaps instead of changing a lot of packages (some image/*
 > > packages also refer to testdata in the parent directory),
 > > we should just invent a fake filesystem for nacl/arm (and
 > > perhaps android/arm.)
 > >
 > > Transferring a big binary program should be faster than
 > > transferring a lot of small files.
 > 
 > We could, but I do like actually using the iOS file system as part
 > of the test.
 > 
 > It doesn't strike me as many packages that use ../testdata (6),
 > though there are one or two others that need related modifications
 > (go/build, as it doesn't have a directory layout).

The other solution is make go_exec_darwin_arm aware of
this and include the testdata from the (direct) parent
directories. for example,

copy:
   compress/testdata
   compress/flate/*
to the application bundle for the compress/flate tests,
and run the program under compress/flate?

We can modify runtime/cgo to chdir to $GOTESTCWD for this.

This approach avoids modifying individual packages for
special arrangements, so it requires less maintenance.
And it should work for a lot of sub-repos too. Unlike the
nacl port, we do want to test subrepos on darwin/arm,
right?

Patch-set: 2