Update patch set 2
Patch Set 2:

> > My new proposal will also fix the following test
 > > failures:
 > > io/ioutil: TestReadDir
 > > path/filepath: TestGlob
 > >
 > > It might fix other tests in path/filepath, but
 > > that depends on whether an app could write to
 > > the root of its bundle.
 > >
 > > path/filepath.TestBug3486 can't be fixed unless
 > > we also copy /test.
 > >
 > > Is there any persistent temporary location on iOS
 > > that we can store a copy of GOROOT? Perhaps that's
 > > the easiest way to fix TestBug3486.
 > 
 > I'm still confused by what is new, besides moving the location of
 > the chdir.

I reread your two proposals. If I understand them correctly:
1. modify the tests to read testdata from the cwd
2. include all of GOROOT.

my proposal is in between, but it doesn't need inclusion of full
GOROOT or modifications to the tests.

For import path a/b/c/d, we include testdata/* and all files
in a, a/b, a/b/c, and, of course, a/b/c/d.

I kinda like the proposal to add -test.cwd to the testing package,
because too many times I build a package test in $GOROOT/src and
then the test fails because it can't access its testdata. Instead
of moving the binary around, I can just rerun with -test.cwd.

Patch-set: 2