Update patch set 2
Patch Set 2:

> We're not adding any new testing command-line flags. The code
 > change in this CL is trivial and test-only and therefore fine with
 > me.

Please note that my proposal doesn't require changing the tests
at all. Adding a -test.cwd is not essential to the proposal.
We can always use an environment variable or some other way to
make the test binary chdir (in the extreme case, a plist property).

My proposal aims to mimic the way how tests are running today, so
we will have the same directory hierarchy on the device, and
we don't need to modify tests that make assumptiosn about the
directory hierarchy.

Changing existing tests to accommodate platform limitations
seems less ideal to me. Besides, the platform limitation is
not that fundamental (unlike nacl).

Also please note that when the darwin/arm64 port goes in, we
will have to redo all the changes for that platform too if
we choose to make modifications to the package tests themselves
rather than to limit the modifications to darwin/arm support
code only (i.e. go_darwin_arm_exec and runtime/cgo.)

Patch-set: 2