Update patch set 2
Patch Set 2:

> > On a second though, perhaps we are doing too much.
 > >
 > > lldb should be able to call functions from the
 > > binary, so we just need to set a breakpoint on
 > > rt0_go, call chdir from lldb, and then continue.
 > >
 > > Then we don't even need the special chdir code
 > > in runtime/cgo. Thoughts?
 > 
 > > On a second though, perhaps we are doing too much.
 > >
 > > lldb should be able to call functions from the
 > > binary, so we just need to set a breakpoint on
 > > rt0_go, call chdir from lldb, and then continue.
 > >
 > > Then we don't even need the special chdir code
 > > in runtime/cgo. Thoughts?
 > 
 > I spent most of today trying to make go_darwin_arm_exec
 > interactively work with lldb via ios-deploy. No go. It simply
 > refuses to pay attention to stdin. I tried an lldb python script to

Perhaps it insists on using a pty. Have you tried providing a pty
to ios-deploy? we should be able to control lldb interactively
that way.
(a simple way is to use script(1) to run the program)

 > get around it, but couldn't make it work.
 > 
 > So to insert a breakpoint to chdir, we would have to modify
 > ios-deploy, which I do not want to touch.
 > 
 > The modifications to the tests for darwin/arm64 should be minimal.
 > I've been thinking about it as I put in these exceptions, and it
 > should only involve changing a line in each relevant test.
 > 
 > I'll give this CL another day, just in case I think of another way
 > around this. The breakpoint chdir is getting close.

Yeah, even better, we break on chdir, and then get the current path out
when runtime/cgo tries to chdir to the application bundle, append
the correct import path (by malloc a new argument and replace the
argument in r0), and then let the process continue. This will leak a
small amount of memory, but should be ok. I do this all the time
with gdb. it should be possible with lldb.

So go_darwin_arm_exec doesn't even need to know the bundle location.

Patch-set: 2