Update patch set 1
Patch Set 1:

I think the likelihood of overflow is significantly smaller with Float than with float64:

- the exponent range is huge (int32)
- the possible precision is huge (uint32, for a theoretical 4B bit mantissa)
- we still have +/-0 for cases where a negative value's mantissa underflows and the value then becomes 0 (e.g. division by very large number)

In IEEE, probably the easiest way to get an Inf is by dividing by 0. Unless it's a programmer error (which we would want to see) this is most likely caused by numbers underflowing to 0 which then are used as divisor.

I've had most of the Inf handling but it still requires handling if NaNs because Inf-Inf = NaN,etc. The point is to avoid panics or extra errors to check in this case. By going directly to NaN (Undef) we remove a little bit of information, but that information is likely of limited use anyway (happy to be convinced otherwise).

I'm also thinking that perhaps we can get rid of -0 for this package, but that may be more tricky to justify.

Patch-set: 1
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0