Update patch set 7
Abandoned

This doesn't work. Anything invoking the external linker can create differing output files, because clang or gcc can insert a random build-id into the file. Will replace this CL with one that just starts invoking new6l etc. If it's happy, we're happy.

I have checked that the linkers produce the same output when external linking is not involved, but it's hard for the go command to predict.

Patch-set: 7
Status: abandoned