debug/dwarf: add support for line tables
THIS IS A DRAFT.  There are no unit tests yet.

This implements a LineReader for line tables that parallels the
existing Reader for debug entries.  It also adds helpers for looking
up line information by PC.

This code is partly based on the debug subrepo's fork of dwarf, but it
is a more complete (and, I believe, correct) implementation of the
spec and has a more general API.  While the debug subrepo's
implementation exposed only a single PC-to-line function, this version
also exposes the line table rows to the caller.  This enables a
broader set of applications; for example, an application that looks up
many PCs may need to create an index for fast lookup.  This
implementation also provides a PC-to-line function, but it is a
utility written entirely in terms of the public API.

Change-Id: Ie157bc817f55e940b6f2e1ae010c5a4e1f29c5c8