Update patch set 3
Patch Set 3: Code-Review+2

The 386 version clearly benefits from the improvements done a little while ago in formatBits. The amd64 version is a bit slower, but I wouldn't worry about it. This is not super speed critical. Instead, if formatBits gets faster, we will automatically get faster here as well which is a net plus.

Thanks.

Patch-set: 3
Label: Code-Review=+2