Update patch set 1
Patch Set 1:

> Are you using the go test -exec option?
 > 
 > I suggest changing the go tool to add that option to the
 > environment when it is used, and having these tests pick it up from
 > the environment.  Then we can use a single mechanism.

Yes, I'm using the -exec option to run the bulk of the tests
(step 3 in the cl description, go_daisy tells how to compile,
go_daisy_exec tells how to run).

The issue is with the (small) subset of the tests that write out
small Go programs and need to do "go build" or "go run". All of
these are typically skipped in a cross-compile setting (see the
top of the test functions modified in this cl, these tests are
being skipped on "android" and "nacl"). As far as I can tell,
none of the existing wrapper mechanisms (-exec, -toolexec,
go_$GOOS_$GOARCH_exec) can satisfactorily cross-compile and
execute these tests remotely.

This new flag provides the extra bit of information to these
tests. For this to work, the tests themselves have to be running
locally (since they need to invoke the Go compiler), but the
nested go build/run commands that they execute happen with the
correct setup for cross-compile and remote-run.

This extra bit of information could be provided by this flag,
or it could be separate flags for 'go_target' and 'go_target_exec',
or it could be provided by an environment variable "${GOTARGET}"
(is this what you're suggesting?).

Patch-set: 1