Update patch set 1
Patch Set 1:

Hmmm, I wasn't thinking clearly about this.

If I understand your approach correctly, I don't like it much.  It means that the tests will fail when running on the target, because they don't have access to a go compiler there.  We shouldn't have tests that we expect to fail.

Most of these tests seem designed to test unusual cases that can't be represented by a simple go test.  They don't care about compiling on the target; they just want to build and run a program under circumstances different than a normal go test.  I think that we should try to think of a way to support that: a way that a go test can arrange to build sources, run them via the -exec option, and collect the output for analysis.  This should be a way that works when cross-compiling.

Patch-set: 1