Update patch set 1
Patch Set 1:

> Hmmm, I wasn't thinking clearly about this.
 > 
 > If I understand your approach correctly, I don't like it much.  It
 > means that the tests will fail when running on the target, because
 > they don't have access to a go compiler there.  We shouldn't have
 > tests that we expect to fail.
 > 
 > Most of these tests seem designed to test unusual cases that can't
 > be represented by a simple go test.  They don't care about
 > compiling on the target; they just want to build and run a program
 > under circumstances different than a normal go test.  I think that
 > we should try to think of a way to support that: a way that a go
 > test can arrange to build sources, run them via the -exec option,
 > and collect the output for analysis.  This should be a way that
 > works when cross-compiling.

This CL is mainly concerned with getting step 4 in the CL description to work, i.e. making it possible to correctly cross-compile and remote-run these tests.

What happens to these tests in step 3 (FAIL or SKIP) is a separate orthogonal concern. In my local test runs, I have replaced the special checks for "android" or "nacl" with a simple check to see if the "go" command is available, and skip the test if it's not. I can send those changes as a separate CL (or include in this one) if you like.

I agree that this solves the problem only for these two packages, but these are the only two in "std" that need this right now ("os" was doing something similar, but we resolved that another way: http://golang.org/cl/4190). If this pattern starts to appear more commonly in "std", it would make sense to add the flag to "go test", plumb it through "testing" (similar to testing.Short()), or provide some functions like testing.GoBuild(...) and testing.GoRun(...) (or new methods on *testing.T) that do the appropriate thing transparently. All that seems like an overkill right now.

Patch-set: 1